Lib/data-parser-master/requirements.txt
appdirs==1.4.3
beautifulsoup4==4.6.0
cssselect==1.0.1
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.5
jieba3k==0.35.1
lxml==3.7.3
newspaper3k==0.1.9
nltk==3.2.4
olefile==0.44
packaging==16.8
Pillow==4.1.1
pyparsing==2.2.0
python-dateutil==2.6.0
PyYAML==3.12
requests==2.14.2
requests-file==1.4.2
six==1.10.0
tldextract==2.0.3
